fix: 이미지 슬라이드 컴포넌트 모달 초기화 현상 해결
